\name{riparia}
\docType{data}
\alias{riparia}
\title{Bank swallow nests}

\description{
2d point pattern depicting the nest locations of Bank Swallows (\emph{Riparia riparia}) in a vertical sand bank. 
}

\usage{riparia}
\format{ppp-object, see package spatstat.}
\source{Photograph of the sand bank taken by H. Högmander in Konnevesi, Finland, 1988.}
\author{
Tuomas Rajala <tuomas.rajala@iki.fi>
}

\references{
Rajala, T. and Penttinen, A. (2012): Bayesian analysis of a Gibbs hard-core point pattern model with varying
repulsion range, \emph{Comp. Stat. Data An.}
}

\keyword{datasets}